moved auth files into auth folders and removed oauth in sign up screen
